Add PixelMapper trait
This PR adds a trait called `PixelMapper` along with its implementations and wrapper struct to enhance the extensibility and organization of the pixel mapping functionality.

- Added a `PixelMapper` trait that defines methods for mapping pixels between an underlying matrix and a visible layout.
- Implemented the `PixelMapper` trait for `MultiplexMapperWrapper` struct. This allows seamless delegation of mapping operations to the underlying `MultiplexMapper`.

We can now pass any `PixelMapper` to `apply_pixel_mapper`. When we are ready to add additional pixel mappers (U-mapper, Rotate, etc), we create another wrapper and implement `PixelMapper` trait for it.